requirements.txt
booleano==1.1a2
Django==4.0.2
djangorestframework==3.13.1
python-decouple==3.6
gunicorn== 20.1.0
mysqlclient==2.1.0
uWSGI>=2.0.18,<2.1
django-mysql==2.2.0
psycopg2>=2.8.6,<2.9